refactor(117): split backend into library and binary
