+++ back/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "picto-back"
+name = "pictou-back"
 version = "0.1.0"
 edition = "2021"
 default-run = "main"
@@ -11,6 +11,10 @@ path = "src/main.rs"
 [[bin]]
 name = "seed"
 path = "src/seed.rs"
+
+[lib]
+name = "pictou"
+path = "src/lib.rs"
 
 [features]
 default = ["testing-incompatible"]
